// frontend/src/api-client/models/TicketRequestDto.ts
export interface TicketRequestDto {
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    url?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    applicationName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    stackTrace?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    device?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    browser?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    resolution?: string | null;
    /**
     * 
     * @type {number}
     * @memberof TicketRequestDto
     */
    priorityId?: number;
    /**
     * 
     * @type {number}
     * @memberof TicketRequestDto
     */
    ticketTypeId?: number;
    /**
     * 
     * @type {number}
     * @memberof TicketRequestDto
     */
    statusId?: number;
    /**
     * 
     * @type {string}
     * @memberof TicketRequestDto
     */
    userOID?: string | null;
    /**
     * 
     * @type {number}
     * @memberof TicketRequestDto
     */
    installedEnvironmentId?: number;
}