// frontend/src/api-client/models/PriorityDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TicketTracking.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PriorityDto
 */
export interface PriorityDto {
    /**
     * 
     * @type {number}
     * @memberof PriorityDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PriorityDto
     */
    title?: string | null;
}

/**
 * Check if a given object implements the PriorityDto interface.
 */
export function instanceOfPriorityDto(value: object): value is PriorityDto {
    return true;
}

export function PriorityDtoFromJSON(json: any): PriorityDto {
    return PriorityDtoFromJSONTyped(json, false);
}

export function PriorityDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PriorityDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
    };
}

export function PriorityDtoToJSON(json: any): PriorityDto {
    return PriorityDtoToJSONTyped(json, false);
}

export function PriorityDtoToJSONTyped(value?: PriorityDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
    };
}