// frontend/src/api-client/models/StatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TicketTracking.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StatusDto
 */
export interface StatusDto {
    /**
     * 
     * @type {number}
     * @memberof StatusDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof StatusDto
     */
    title?: string | null;
}

/**
 * Check if a given object implements the StatusDto interface.
 */
export function instanceOfStatusDto(value: object): value is StatusDto {
    return true;
}

export function StatusDtoFromJSON(json: any): StatusDto {
    return StatusDtoFromJSONTyped(json, false);
}

export function StatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StatusDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
    };
}

export function StatusDtoToJSON(json: any): StatusDto {
    return StatusDtoToJSONTyped(json, false);
}

export function StatusDtoToJSONTyped(value?: StatusDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
    };
}