// frontend/src/api-client/models/TicketTypeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TicketTracking.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TicketTypeDto
 */
export interface TicketTypeDto {
    /**
     * 
     * @type {number}
     * @memberof TicketTypeDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof TicketTypeDto
     */
    title?: string | null;
}

/**
 * Check if a given object implements the TicketTypeDto interface.
 */
export function instanceOfTicketTypeDto(value: object): value is TicketTypeDto {
    return true;
}

export function TicketTypeDtoFromJSON(json: any): TicketTypeDto {
    return TicketTypeDtoFromJSONTyped(json, false);
}

export function TicketTypeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketTypeDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
    };
}

export function TicketTypeDtoToJSON(json: any): TicketTypeDto {
    return TicketTypeDtoToJSONTyped(json, false);
}

export function TicketTypeDtoToJSONTyped(value?: TicketTypeDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
    };
}